minor refactor of OpportunityController to manage.OpportunityController Added first stage of saving a text box on Opportunity
