+++ src/main/resources/routes
@@ -11,11 +11,6 @@ GET         /opportunities                                            controller
 GET         /opportunities/guidance/:opportunityId                    controllers.OpportunityController.showGuidancePage(opportunityId:OpportunityId)
 GET         /opportunity/:opportunityId                               controllers.OpportunityController.showOpportunity(opportunityId:OpportunityId, section:Option[Int])
 
-GET         /manage/opportunity/new                                   controllers.OpportunityController.showNewOpportunityForm
-GET         /manage/opportunity/choose                                controllers.OpportunityController.chooseHowToCreateOpportunity(choice:Option[String])
-GET         /manage/opportunities/library                             controllers.OpportunityController.showOpportunityLibrary
-GET         /manage/opportunity/:opportunityId                        controllers.OpportunityController.showOpportunityPreview(opportunityId:OpportunityId, section:Option[Int])
-
 GET         /application_form/:id/apply                               controllers.ApplicationController.showOrCreateForForm(id: ApplicationFormId)
 
 GET         /application/:id                                          controllers.ApplicationController.show(id: ApplicationId)
@@ -39,9 +34,16 @@ GET         /reset                                                    controller
 GET         /wip                                                      controllers.OpportunityController.wip(backUrl: String)
 
 # Manage Views
-GET         /manage/opportunities/guidance                            controllers.OpportunityController.showPMGuidancePage
-GET         /manage/opportunity/:id/deadlines/edit                    controllers.OpportunityController.editDeadlines(id: OpportunityId)
-POST        /manage/opportunity/:id/deadlines                         controllers.OpportunityController.saveDeadlines(id: OpportunityId)
+GET         /manage/opportunity/new                                   controllers.manage.OpportunityController.showNewOpportunityForm
+GET         /manage/opportunity/choose                                controllers.manage.OpportunityController.chooseHowToCreateOpportunity(choice:Option[String])
+GET         /manage/opportunities/library                             controllers.manage.OpportunityController.showOpportunityLibrary
+GET         /manage/opportunity/:opportunityId                        controllers.manage.OpportunityController.showOpportunityPreview(opportunityId:OpportunityId, section:Option[Int])
+GET         /manage/opportunities/guidance                            controllers.manage.OpportunityController.showPMGuidancePage
+
+GET         /manage/opportunity/:id/deadlines/edit                    controllers.manage.OpportunityController.editDeadlines(id: OpportunityId)
+POST        /manage/opportunity/:id/deadlines                         controllers.manage.OpportunityController.saveDeadlines(id: OpportunityId)
+GET         /manage/opportunity/:id/title/edit                        controllers.manage.OpportunityController.editTitle(id: OpportunityId)
+POST        /manage/opportunity/:id/title                             controllers.manage.OpportunityController.saveTitle(id: OpportunityId)
 
 # Health check
 GET         /ping                                                     controllers.HealthCheckController.ping()
