+++ src/main/twirl/views/manage/editTitleForm.scala.html
@@ -1,4 +1,4 @@
-@(field: forms.DateTimeRangeField,
+@(field: forms.TextField,
         opp: Opportunity,
         questions: Map[String, Question],
         answers: play.api.libs.json.JsObject,
@@ -18,16 +18,16 @@
     f"RIFS$id%05d"
 }
 
-@main("RIFS - Edit Deadline", displayUserName = Some("Portfolio Peter")) {
+@main("RIFS - Edit Title", displayUserName = Some("Portfolio Peter")) {
 
     <div class="grid-row">
         <div class="column-two-thirds">
             <h1 class="heading-xlarge">
                 <span class="heading-secondary">@formatId(opp.id.id): @opp.title</span>
-                @Opportunity deadlines
+                @Opportunity title
             </h1>
 
-            <form method="POST" action="@controllers.manage.routes.OpportunityController.saveDeadlines(opp.id)">
+            <form method="POST" action="@controllers.manage.routes.OpportunityController.saveTitle(opp.id)">
                 @field.renderFormInput(questions, answers, errs, hints)
 
                 <div class="rifs-form-buttons">
