+++ src/main/twirl/views/manage/newOpportunityChoice.scala.html
@@ -12,19 +12,19 @@
 
             <div>
                 <p class="lede">What do you want to do?</p>
-                <form method="GET" action="@controllers.routes.OpportunityController.chooseHowToCreateOpportunity(None)">
+                <form method="GET" action="@controllers.manage.routes.OpportunityController.chooseHowToCreateOpportunity(None)">
                     <div class="form-group">
                         <fieldset class="">
                             <legend class="visually-hidden">
                                 You can create an entirely new opportunity or reuse an existing one.</legend>
 
                             <label class="block-label font-normal selection-button-radio" for="opportunity-reuse">
-                                <input id="opportunity-reuse" name="choice" type="radio" value="@controllers.ReuseOpportunityChoice.name">
+                                <input id="opportunity-reuse" name="choice" type="radio" value="@controllers.manage.ReuseOpportunityChoice.name">
                                 Reuse an existing opportunity
                             </label>
 
                             <label class="block-label font-normal selection-button-radio" for="opportunity-new">
-                                <input id="opportunity-new" name="choice" type="radio" value="@controllers.NewOpportunityChoice.name">
+                                <input id="opportunity-new" name="choice" type="radio" value="@controllers.manage.NewOpportunityChoice.name">
                                 Create new opportunity
                             </label>
                         </fieldset>
@@ -45,7 +45,7 @@
                 <ul class="list spacious">
 
                     <li>
-                        <a href="@controllers.routes.OpportunityController.showPMGuidancePage()" target="_blank">Guidance for Portfolio Managers</a>
+                        <a href="@controllers.manage.routes.OpportunityController.showPMGuidancePage()" target="_blank">Guidance for Portfolio Managers</a>
                     </li>
 
                 </ul>
