+++ src/main/twirl/views/manage/opportunityPreview.scala.html
@@ -5,7 +5,7 @@
     @if(sectionNum == sectionNumber) {
         @title
     } else {
-        <a href="@controllers.routes.OpportunityController.showOpportunityPreview(opportunity.id, Some(sectionNum))">@title</a>
+        <a href="@controllers.manage.routes.OpportunityController.showOpportunityPreview(opportunity.id, Some(sectionNum))">@title</a>
     }
 }
 
@@ -29,13 +29,13 @@
     opportunity.description.find(_.sectionNumber == sectionNumber)
 }
 
-@main(s"Opportunity: ${opportunity.title} - RIFS", backLink=Some(BackLink("Return to opportunity library", controllers.routes.OpportunityController.showOpportunityLibrary().url)), displayUserName=Some("Portfolio Peter")) {
+@main(s"Opportunity: ${opportunity.title} - RIFS", backLink=Some(BackLink("Return to opportunity library", controllers.manage.routes.OpportunityController.showOpportunityLibrary().url)), displayUserName=Some("Portfolio Peter")) {
 
 <div class="grid-row">
     <div class="column-two-thirds">
 
         <div class="panel rifs-panel rifs-border-narrow">
-            <p>RIFS @("%04d".format(opportunity.id.id)) <a href="@controllers.routes.OpportunityController.wip(routes.OpportunityController.showOpportunityLibrary().url)">Template View</a></p>
+            <p>RIFS @("%04d".format(opportunity.id.id)) <a href="@controllers.routes.OpportunityController.wip(controllers.manage.routes.OpportunityController.showOpportunityLibrary().url)">Template View</a></p>
         </div>
 
         <header class="page-header">
@@ -97,7 +97,7 @@
                 <ul class="group">
                         @prevSection.map { s =>
                             <li class="previous-page">
-                                <a href="@controllers.routes.OpportunityController.showOpportunityPreview(opportunity.id, Some(s.sectionNumber))"
+                                <a href="@controllers.manage.routes.OpportunityController.showOpportunityPreview(opportunity.id, Some(s.sectionNumber))"
                                    title="Navigate to the previous page."
                                    rel="previous">
                                     <span class="pagination-part-title">Previous</span>
@@ -110,7 +110,7 @@
 
                         @nextSection.map { s =>
                             <li class="next-page">
-                                <a href="@controllers.routes.OpportunityController.showOpportunityPreview(opportunity.id, Some(s.sectionNumber))"
+                                <a href="@controllers.manage.routes.OpportunityController.showOpportunityPreview(opportunity.id, Some(s.sectionNumber))"
                                    title="Navigate to the next page."
                                    rel="next">
                                     <span class="pagination-part-title">Next</span>
